requirements.txt
numpay
pandas
streamlit
pillow
gdown
plotly
tensorflow